backend/requirements.txt

fastapi==0.109.2
uvicorn==0.27.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
ultralytics==8.1.18
python-dotenv==1.0.1
websockets==12.0
